requirements.txt
python-telegram-bot
emoji
requests